BSIS-2608 Change counsellingStatus type in search endpoint
The frontend is sending the id instead of the enum for counsellingStatus. Changed the type to Long and the way it's used in the query.

The current view model had to be changed to as it was expecting an int instead of a long.

+++ src/main/java/org/jembi/bsis/backingform/PostDonationCounsellingBackingForm.java
@@ -32,7 +32,7 @@ public class PostDonationCounsellingBackingForm {
     return counsellingStatus;
   }
 
-  public void setCounsellingStatus(Integer counsellingStatusId) {
+  public void setCounsellingStatus(Long counsellingStatusId) {
     if (counsellingStatusId == null) {
       return;
     }
